requirement.txt
numpy	
opencv-python 
scikit-learn 
joblib	
Pillow (PIL)
streamlit
matplotlib
seaborn